[update] divite exe_inst into module of extension
