+++ src/cpu/execution.rs
@@ -3,8 +3,6 @@ mod inst_32;
 
 use super::CPU;
 use super::instruction::{Instruction, Extensions};
-use inst_16::exe_cinst;
-use inst_32::exe_inst;
 
 pub trait Execution {
     fn execution(&self, cpu: &mut CPU);
@@ -15,8 +13,8 @@ impl Execution for Instruction {
         dbg!(self);
 
         match self.opc_to_extension() {
-            Extensions::C => exe_cinst(self, cpu),
-            _ => exe_inst(self, cpu),
+            Extensions::C => inst_16::exe_cinst(self, cpu),
+            _ => inst_32::exe_inst(self, cpu),
         }
 
         cpu.regs.show();
