+++ src/cpu/execution/inst_32/m_extension.rs
@@ -1,0 +1,9 @@
+use crate::cpu::CPU;
+use crate::cpu::instruction::Instruction;
+
+pub fn exec(inst: &Instruction, _cpu: &mut CPU) {
+    match inst.opc {
+        _ => panic!("M extension is not implemented."),
+    }
+}
+
